Split out pricing map to dedicated file
